+++ lib/AletheiaCore/test/runtests.jl
@@ -1,4 +1,5 @@
 using Test
+using Serialization
 using Random
 using Aqua
 using JET